requirements.txt
backtesting>=0.3.3
pandas>=1.3.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=0.19.0
pandas_ta
